auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
PyJWT
python-multipart
slowapi 